refactor: Move route handlers under runtime routes package
