+++ internal/runtime/routes/handler/chatCompletions/stream.go
