+++ internal/runtime/routes/handler/sendSSE.go
